Extraacts PhotoSubmissionConfirmation
